# sarc/docs/swagger.yaml
definitions:
...
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  domain.Reservation:
    properties:
      lectureId:
        type: integer
      observation:
        type: string
      reservationId:
        type: integer
      resources:
        items:
          $ref: '#/definitions/domain.Resource'
        type: array
    type: object
  domain.Resource:
    properties:
      characteristics:
        description: |-
          Characteristics is an array of strings stored as Postgres text[].
          For Swagger, treat as []string.
        items:
          type: string
        type: array
      description:
        type: string
      resourceId:
        type: integer
      resourceTypeId:
        type: integer
      status:
        $ref: '#/definitions/domain.ResourceStatus'
    type: object
  domain.ResourceStatus:
    enum:
    - available
    - unavailable
    - reserved
    type: string
    x-enum-varnames:
    - ResourceStatusAvailable
    - ResourceStatusUnavailable
    - ResourceStatusReserved
...
info:
  contact: {}
paths:
...
  /reservations:
    get:
      description: Retrieves all reservations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Reservation'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: Creates a new reservation in the system
      parameters:
      - description: Reservation data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/domain.Reservation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Create a new reservation
      tags:
      - reservations
  /reservations/{id}:
    delete:
      description: Deletes a reservation by its ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Delete a reservation
      tags:
      - reservations
    get:
      description: Retrieves a reservation by its ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get reservation by ID
      tags:
      - reservations
    put:
      consumes:
      - application/json
      description: Updates the reservation information for the given reservation ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reservation data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/domain.Reservation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Invalid ID or bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update an existing reservation
      tags:
      - reservations
...
swagger: "2.0"